Implement core game modules with crafting, inventory, and UI components
