+++ factorio-game/src/components/GameIcon.tsx
