+++ factorio-game/src/components/ItemCard.tsx
