+++ factorio-game/src/components/ItemDetailDialog.tsx
