+++ factorio-game/src/components/ProductionStats.tsx
