+++ factorio-game/src/components/QuickCraftButtons.tsx
