+++ factorio-game/src/components/TechnologyModule.tsx
