+++ factorio-game/src/utils/format.ts
