--- public/data/1.1/categories.json
@@ -1,44 +1,0 @@
-[
-  {
-    "id": "logistics",
-    "name": "物流",
-    "icon": "logistics",
-    "order": 1
-  },
-  {
-    "id": "production",
-    "name": "生产",
-    "icon": "production",
-    "order": 2
-  },
-  {
-    "id": "intermediate-product",
-    "name": "中间产品",
-    "icon": "intermediate",
-    "order": 3
-  },
-  {
-    "id": "combat",
-    "name": "战斗",
-    "icon": "combat",
-    "order": 4
-  },
-  {
-    "id": "fluid",
-    "name": "流体",
-    "icon": "fluid",
-    "order": 5
-  },
-  {
-    "id": "science-pack",
-    "name": "科技",
-    "icon": "science",
-    "order": 6
-  },
-  {
-    "id": "raw-resource",
-    "name": "原材料",
-    "icon": "raw-resource",
-    "order": 7
-  }
-]
