--- public/data/1.1/data.json
@@ -1,10 +1,0 @@
-{
-  "version": "1.1",
-  "hash": "test-hash-v1",
-  "files": {
-    "categories": "categories.json",
-    "items": "items.json",
-    "recipes": "recipes.json",
-    "technologies": "technologies.json"
-  }
-}
