--- public/data/1.1/items.json
@@ -1,212 +1,0 @@
-[
-  {
-    "id": "iron-ore",
-    "name": "铁矿石",
-    "category": "raw-resource",
-    "stack_size": 50,
-    "icon": "iron-ore",
-    "unlocked": true
-  },
-  {
-    "id": "copper-ore",
-    "name": "铜矿石",
-    "category": "raw-resource",
-    "stack_size": 50,
-    "icon": "copper-ore",
-    "unlocked": true
-  },
-  {
-    "id": "coal",
-    "name": "煤炭",
-    "category": "raw-resource",
-    "stack_size": 50,
-    "fuel_value": "4MJ",
-    "icon": "coal",
-    "unlocked": true
-  },
-  {
-    "id": "stone",
-    "name": "石头",
-    "category": "raw-resource",
-    "stack_size": 50,
-    "icon": "stone",
-    "unlocked": true
-  },
-  {
-    "id": "wood",
-    "name": "木材",
-    "category": "raw-resource",
-    "stack_size": 100,
-    "fuel_value": "2MJ",
-    "icon": "wood",
-    "unlocked": true
-  },
-  {
-    "id": "water",
-    "name": "水",
-    "category": "fluid",
-    "type": "fluid",
-    "icon": "water",
-    "unlocked": true
-  },
-  {
-    "id": "steam",
-    "name": "蒸汽",
-    "category": "fluid",
-    "type": "fluid",
-    "icon": "steam",
-    "unlocked": true
-  },
-  {
-    "id": "iron-plate",
-    "name": "铁板",
-    "category": "intermediate-product",
-    "stack_size": 100,
-    "icon": "iron-plate",
-    "unlocked": true
-  },
-  {
-    "id": "copper-plate",
-    "name": "铜板",
-    "category": "intermediate-product",
-    "stack_size": 100,
-    "icon": "copper-plate",
-    "unlocked": true
-  },
-  {
-    "id": "steel-plate",
-    "name": "钢板",
-    "category": "intermediate-product",
-    "stack_size": 100,
-    "icon": "steel-plate",
-    "unlocked": true
-  },
-  {
-    "id": "iron-gear-wheel",
-    "name": "铁齿轮",
-    "category": "intermediate-product",
-    "stack_size": 100,
-    "icon": "iron-gear-wheel",
-    "unlocked": true
-  },
-  {
-    "id": "electronic-circuit",
-    "name": "电路板",
-    "category": "intermediate-product",
-    "stack_size": 200,
-    "icon": "electronic-circuit",
-    "unlocked": true
-  },
-  {
-    "id": "automation-science-pack",
-    "name": "自动化科学包",
-    "category": "science-pack",
-    "stack_size": 200,
-    "icon": "automation-science-pack",
-    "unlocked": true
-  },
-  {
-    "id": "burner-mining-drill",
-    "name": "热力采掘机",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "burner-mining-drill",
-    "unlocked": true
-  },
-  {
-    "id": "electric-mining-drill",
-    "name": "电力采掘机",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "electric-mining-drill",
-    "unlocked": true
-  },
-  {
-    "id": "stone-furnace",
-    "name": "石炉",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "stone-furnace",
-    "unlocked": true
-  },
-  {
-    "id": "electric-furnace",
-    "name": "电炉",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "electric-furnace",
-    "unlocked": false
-  },
-  {
-    "id": "assembling-machine-1",
-    "name": "装配机1型",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "assembling-machine-1",
-    "unlocked": true
-  },
-  {
-    "id": "assembling-machine-2",
-    "name": "装配机2型",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "assembling-machine-2",
-    "unlocked": false
-  },
-  {
-    "id": "offshore-pump",
-    "name": "海水泵",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "offshore-pump",
-    "unlocked": true
-  },
-  {
-    "id": "boiler",
-    "name": "锅炉",
-    "category": "production",
-    "stack_size": 50,
-    "icon": "boiler",
-    "unlocked": true
-  },
-  {
-    "id": "steam-engine",
-    "name": "蒸汽机",
-    "category": "production",
-    "stack_size": 10,
-    "icon": "steam-engine",
-    "unlocked": true
-  },
-  {
-    "id": "lab",
-    "name": "研究实验室",
-    "category": "production",
-    "stack_size": 10,
-    "icon": "lab",
-    "unlocked": true
-  },
-  {
-    "id": "copper-cable",
-    "name": "铜线",
-    "category": "intermediate-product",
-    "stack_size": 200,
-    "icon": "copper-cable",
-    "unlocked": true
-  },
-  {
-    "id": "pipe",
-    "name": "管道",
-    "category": "logistics",
-    "stack_size": 100,
-    "icon": "pipe",
-    "unlocked": true
-  },
-  {
-    "id": "transport-belt",
-    "name": "传送带",
-    "category": "logistics",
-    "stack_size": 100,
-    "icon": "transport-belt",
-    "unlocked": true
-  }
-]
