--- public/data/1.1/technologies.json
@@ -1,81 +1,0 @@
-[
-  {
-    "id": "automation",
-    "name": "自动化",
-    "description": "解锁装配机1型",
-    "icon": "automation",
-    "prerequisites": [],
-    "effects": [
-      {
-        "type": "unlock-recipe",
-        "recipe": "assembling-machine-1"
-      }
-    ],
-    "unit": {
-      "count": 10,
-      "time": 10,
-      "ingredients": {
-        "automation-science-pack": 1
-      }
-    },
-    "order": "a-b-a"
-  },
-  {
-    "id": "electronics",
-    "name": "电子学",
-    "description": "解锁电路板制作",
-    "icon": "electronics",
-    "prerequisites": ["automation"],
-    "effects": [
-      {
-        "type": "unlock-recipe",
-        "recipe": "electronic-circuit"
-      }
-    ],
-    "unit": {
-      "count": 30,
-      "time": 15,
-      "ingredients": {
-        "automation-science-pack": 1
-      }
-    },
-    "order": "a-d-a"
-  },
-  {
-    "id": "electric-energy-distribution-1",
-    "name": "电力分配1",
-    "description": "解锁中型电线杆和地下管道",
-    "icon": "electric-energy-distribution-1",
-    "prerequisites": ["electronics"],
-    "effects": [],
-    "unit": {
-      "count": 30,
-      "time": 30,
-      "ingredients": {
-        "automation-science-pack": 1
-      }
-    },
-    "order": "c-e-a"
-  },
-  {
-    "id": "advanced-material-processing",
-    "name": "高级材料加工",
-    "description": "解锁钢铁熔炉",
-    "icon": "advanced-material-processing",
-    "prerequisites": [],
-    "effects": [
-      {
-        "type": "unlock-recipe",
-        "recipe": "steel-furnace"
-      }
-    ],
-    "unit": {
-      "count": 75,
-      "time": 30,
-      "ingredients": {
-        "automation-science-pack": 1
-      }
-    },
-    "order": "c-c-a"
-  }
-]
